fix(sdk): stabilize simulate trace diagnostics
